Create Forsaken LMS Changer
This changes the LMS in forsaken. A lot of the lms in forsaken sounds super boring but u can change that by uploading a singular song on roblox studio and then copying the asset ID and pasting it into the ID section of the script. Please be aware I am still testing the script to successfully post this without bugs.